feat: CatToyService와 CatToyService 분리
